backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
pydantic==1.10.12
